Fixing typo for the profile name to have proper 's suffix https://bugster.forgerock.org/jira/browse/OPENIDM-1068
Removing path prefixes and letting it up to the configuration to specify full or relative paths

http://sources.forgerock.org/cru/CR-1168
https://bugster.forgerock.org/jira/browse/OPENIDM-1100

http://sources.forgerock.org/cru/CR-1157 - Commons UI: Updating patch actions to try using add when replace fails

These changes allow for the possibility of a non-existent password property on the managed/user object as a valid state.  This could arise when, for example, a record was create via recon from a source system which doesn't expose a cleartext password.  Code changes to policy and onCreate script to make this more sensible, and within the UI code to do an "add"-type of patch rather than a "replace" patch (which would fail, since the property doesn't exist).

git-svn-id: https://svn.forgerock.org/openidm/branches/2.1.x@1888 d98387aa-ee2c-4292-a9e6-504d2a719fd3